refactor: replace custom SuppressOutput implementation with import from hssm.utils
